// fron/cliente/src/app/models/Edm01.ts
export class Edm01titulo{
    _id?: number;
    edm01titulo1: string;
    edm01titulo2: string;
    edm01titulo3: string;
    edm01titulo4: string;
    edm01titulo5: string; 
    edm01titulo6: string;
    edm01titulo7: string;
    edm01titulo8: string;
    edm01titulo9: string;
    edm01titulo10: string;
    edm01titulo11: string;
    edm01titulo12: string;
    edm01titulo13: string;
    edm01titulo14: string;
    edm01titulo15:string;
    edm01titulo16: string;
    edm01titulo17: string;
    edm01titulo18: string;
    edm01titulo19: string;
    edm01titulo20: string;
    edm01titulo21: string;
    edm01titulo22: string;
    edm01titulo23: string;
    edm01titulo24: string;
    edm01titulo25: string;
    edm01titulo26: string;
    

    constructor(edm01titulo1:string,edm01titulo2:string,edm01titulo3:string,edm01titulo4:string ,edm01titulo5: string,
       edm01titulo6: string, edm01titulo7: string, edm01titulo8: string, edm01titulo9: string, edm01titulo10: string, 
       edm01titulo11: string, edm01titulo12: string, edm01titulo13: string, edm01titulo14: string, edm01titulo15: string,
       edm01titulo16: string, edm01titulo17: string, edm01titulo18: string, edm01titulo19: string, edm01titulo20: string,
       edm01titulo21: string, edm01titulo22: string, edm01titulo23: string, edm01titulo24: string, edm01titulo25: string,
       edm01titulo26: string,
       ){
        this.edm01titulo1=edm01titulo1;
        this.edm01titulo2=edm01titulo2;
        this.edm01titulo3=edm01titulo3;
        this.edm01titulo4=edm01titulo4;
        this.edm01titulo5=edm01titulo5;
        this.edm01titulo6=edm01titulo6;
        this.edm01titulo7=edm01titulo7;
        this.edm01titulo8=edm01titulo8;
        this.edm01titulo9=edm01titulo9;
        this.edm01titulo10=edm01titulo10;
        this.edm01titulo11=edm01titulo11;
        this.edm01titulo12=edm01titulo12;
        this.edm01titulo13=edm01titulo13;
        this.edm01titulo14=edm01titulo14;
        this.edm01titulo15=edm01titulo15;
        this.edm01titulo16=edm01titulo16;
        this.edm01titulo17=edm01titulo17;
        this.edm01titulo18=edm01titulo18;
        this.edm01titulo19=edm01titulo19;
        this.edm01titulo20=edm01titulo20;
        this.edm01titulo21=edm01titulo21;
        this.edm01titulo22=edm01titulo22;
        this.edm01titulo23=edm01titulo23;
        this.edm01titulo24=edm01titulo24;
        this.edm01titulo25=edm01titulo25;
        this.edm01titulo26=edm01titulo26;

    }

}